Add authorisation to api routes

+++ server/lib/components/routes/accounts.js
@@ -6,9 +6,11 @@ export default function(options = {}) {
   function start({ pkg, app, loggerMiddleware, store, }, cb) {
 
     app.get('/api/accounts', async (req, res, next) => {
-      const limit = req.query.limit ? parseInt(req.query.limit, 10) : undefined;
-      const offset = req.query.offset ? parseInt(req.query.offset, 10) : undefined;
       try {
+        if (!req.user.hasPermission('placeholder', 'accounts-read')) return next(Boom.forbidden());
+
+        const limit = req.query.limit ? parseInt(req.query.limit, 10) : undefined;
+        const offset = req.query.offset ? parseInt(req.query.offset, 10) : undefined;
         const accounts = await store.listAccounts(limit, offset);
         res.json(accounts);
       } catch (err) {
@@ -18,6 +20,8 @@ export default function(options = {}) {
 
     app.get('/api/accounts/:id', async (req, res, next) => {
       try {
+        if (!req.user.hasPermission('placeholder', 'accounts-read')) return next(Boom.forbidden());
+
         const account = await store.getAccount(req.params.id);
         return account ? res.json(account) : next();
       } catch (err) {
@@ -26,9 +30,11 @@ export default function(options = {}) {
     });
 
     app.post('/api/accounts', bodyParser.json(), async (req, res, next) => {
-      if (!req.body.displayName) return next(Boom.badRequest('displayName is required'));
-
       try {
+        if (!req.user.hasPermission('placeholder', 'accounts-write')) return next(Boom.forbidden());
+
+        if (!req.body.displayName) return next(Boom.badRequest('displayName is required'));
+
         const data = { displayName: req.body.displayName, };
         const meta = { date: new Date(), account: req.user.id, };
         const account = await store.saveAccount(data, meta);
@@ -40,6 +46,8 @@ export default function(options = {}) {
 
     app.delete('/api/accounts/:id', async (req, res, next) => {
       try {
+        if (!req.user.hasPermission('placeholder', 'accounts-write')) return next(Boom.forbidden());
+
         await store.deleteAccount(req.params.id, { date: new Date(), account: req.user.id, });
         res.status(204).send();
       } catch (err) {
@@ -48,12 +56,14 @@ export default function(options = {}) {
     });
 
     app.post('/api/identities', bodyParser.json(), async (req, res, next) => {
-      if (!req.body.account) return next(Boom.badRequest('account is required'));
-      if (!req.body.name) return next(Boom.badRequest('name is required'));
-      if (!req.body.provider) return next(Boom.badRequest('provider is required'));
-      if (!req.body.type) return next(Boom.badRequest('type is required'));
-
       try {
+        if (!req.user.hasPermission('placeholder', 'accounts-write')) return next(Boom.forbidden());
+
+        if (!req.body.account) return next(Boom.badRequest('account is required'));
+        if (!req.body.name) return next(Boom.badRequest('name is required'));
+        if (!req.body.provider) return next(Boom.badRequest('provider is required'));
+        if (!req.body.type) return next(Boom.badRequest('type is required'));
+
         const data = { name: req.body.name, provider: req.body.provider, type: req.body.type, };
         const meta = { date: new Date(), account: req.user.id, };
         const identity = await store.saveIdentity(req.body.account, data, meta);
@@ -65,6 +75,8 @@ export default function(options = {}) {
 
     app.delete('/api/identities/:id', async (req, res, next) => {
       try {
+        if (!req.user.hasPermission('placeholder', 'accounts-write')) return next(Boom.forbidden());
+
         await store.deleteIdentity(req.params.id, { date: new Date(), account: req.user.id, });
         res.status(204).send();
       } catch (err) {
@@ -73,6 +85,9 @@ export default function(options = {}) {
     });
 
     app.post('/api/roles', bodyParser.json(), async (req, res, next) => {
+
+      if (!req.user.hasPermission('placeholder', 'accounts-write')) return next(Boom.forbidden());
+
       if (!req.body.account) return next(Boom.badRequest('account is required'));
       if (!req.body.role) return next(Boom.badRequest('role is required'));
 
@@ -86,6 +101,9 @@ export default function(options = {}) {
     });
 
     app.delete('/api/roles/:id', async (req, res, next) => {
+
+      if (!req.user.hasPermission('placeholder', 'accounts-write')) return next(Boom.forbidden());
+
       try {
         await store.revokeRole(req.params.id, { date: new Date(), account: req.user.id, });
         res.status(204).send();
